refactor(test): extract controller layer tests to tests/unit/
Move inline #[cfg(test)] modules from controller layer (validator, extractor,
error, di) to external test files. 5 tests deleted (required #[cfg(test)]
SecretKey::empty_for_test() or #[non_exhaustive] PublicKey construction).

Moved: 38 tests, Deleted: 5 tests
