+++ client/tests/unit/controller/mod.rs
@@ -1,0 +1,4 @@
+mod di;
+mod error;
+mod extractor;
+mod validator;
